pyproject.toml
[project]
name = "magnetbank"
version = "0.1.0"
description = "MagnetBank Monorepo"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anyio>=4.13.0",
    "apscheduler>=3.11.2",
    "beem>=0.24.26",
    "bencode-py>=4.0.0",
    "beautifulsoup4>=4.14.3",
    "blinker>=1.9.0",
    "cachelib>=0.13.0",
    "click>=8.3.3",
    "click-shell>=2.1",
    "flask>=3.1.3",
    "flask-paginate>=2024.4.12",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "idna>=3.13",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "lxml>=6.1.0",
    "markdown>=3.10.2",
    "markupsafe>=3.0.3",
    "pymongo>=4.17.0",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "requests>=2.33.1",
    "ruamel-yaml>=0.19.1",
    "scrypt>=0.9.4",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.49",
    "tzdata>=2026.2",
    "tzlocal>=5.3.1",
    "urllib3>=2.6.3",
    "werkzeug>=3.1.8",
    "websocket-client>=1.9.0",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[tool.uv]
managed = true

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.black]
line-length = 88
target-version = ['py312']
